refactor: adjust types so that react renderer exports don't required extra generic parameter
