+++ src/core/index.ts
@@ -1,10 +1,9 @@
-import { CreateRenderer, Renderer, RenderResult, RenderHook, RenderHookOptions } from '../types'
-import { ResultContainer } from '../types/internal'
+import { CreateRenderer, Renderer, RenderResult, RenderHookOptions } from '../types'
 
 import { asyncUtils } from './asyncUtils'
 import { cleanup, addCleanup, removeCleanup } from './cleanup'
 
-function resultContainer<TValue>(): ResultContainer<TValue> {
+function resultContainer<TValue>() {
   const results: Array<{ value?: TValue; error?: Error }> = []
   const resolvers: Array<() => void> = []
 
@@ -43,12 +42,12 @@ function resultContainer<TValue>(): ResultContainer<TValue> {
 function createRenderHook<
   TProps,
   TResult,
-  TOptions extends object,
+  TRendererOptions extends object,
   TRenderer extends Renderer<TProps>
->(createRenderer: CreateRenderer<TProps, TResult, TOptions, TRenderer>) {
-  const renderHook: RenderHook<TProps, TResult, TOptions, TRenderer> = (
-    callback,
-    options = {} as RenderHookOptions<TProps, TOptions>
+>(createRenderer: CreateRenderer<TProps, TResult, TRendererOptions, TRenderer>) {
+  const renderHook = (
+    callback: (props: TProps) => TResult,
+    options = {} as RenderHookOptions<TProps> & TRendererOptions
   ) => {
     const { result, setValue, setError, addResolver } = resultContainer<TResult>()
     const renderProps = { callback, setValue, setError }
