+++ src/server/pure.ts
@@ -2,10 +2,9 @@ import ReactDOMServer from 'react-dom/server'
 import ReactDOM from 'react-dom'
 import { act } from 'react-dom/test-utils'
 
-import { RendererProps } from '../types'
-import { RendererOptions } from '../types/react'
+import { RendererProps, RendererOptions } from '../types/react'
 
-import { createRenderHook, cleanup, addCleanup, removeCleanup } from '../core'
+import { createRenderHook } from '../core'
 import { createTestHarness } from '../helpers/createTestHarness'
 
 function createServerRenderer<TProps, TResult>(
@@ -60,7 +59,8 @@ function createServerRenderer<TProps, TResult>(
 
 const renderHook = createRenderHook(createServerRenderer)
 
-export { renderHook, act, cleanup, addCleanup, removeCleanup }
+export { renderHook, act }
 
-export * from '../types'
+export { cleanup, addCleanup, removeCleanup } from '../core'
+
 export * from '../types/react'
