--- src/types/internal.ts
@@ -1,8 +1,0 @@
-import { RenderResult } from '.'
-
-export type ResultContainer<TValue> = {
-  result: RenderResult<TValue>
-  addResolver: (resolver: () => void) => void
-  setValue: (val: TValue) => void
-  setError: (error: Error) => void
-}
