+++ src/types/react.ts
@@ -1,6 +1,6 @@
 import { ComponentType } from 'react'
 
-import { RenderHookOptions, RenderHookResult, Act } from '.'
+import { RenderHookOptions as BaseRenderHookOptions, RenderHookResult, Act } from '.'
 
 export type WrapperComponent<TProps> = ComponentType<TProps>
 
@@ -8,13 +8,19 @@ export type RendererOptions<TProps> = {
   wrapper?: WrapperComponent<TProps>
 }
 
+export type RenderHookOptions<TProps> = BaseRenderHookOptions<TProps> & {
+  wrapper?: WrapperComponent<TProps>
+}
+
 export type ReactHooksRenderer = {
   renderHook: <TProps, TResult>(
     callback: (props: TProps) => TResult,
-    options?: RenderHookOptions<TProps, RendererOptions<TProps>>
+    options?: RenderHookOptions<TProps>
   ) => RenderHookResult<TProps, TResult>
   act: Act
   cleanup: () => void
   addCleanup: (callback: () => Promise<void> | void) => () => void
   removeCleanup: (callback: () => Promise<void> | void) => void
 }
+
+export * from '.'
